Vita real-hardware bring-up: logging, stack/heap, printf %z, diagnostics
port_log: write the log file through a raw fd + write(2) instead of stdio.
The buffered path stalled indefinitely on real hardware the first time
libc's internal 1024-byte buffer filled, which looked exactly like a hang.
Keeps the two independent writer threads (sceClibPrintf vs file) so neither
can block the other or the caller, and gives them explicit stack sizes.

port.cpp: define sceUserMainThreadStackSize (4MiB) and _newlib_heap_size_user
(256MiB). Neither was ever set - a coredump showed the crashing thread's
whole stack mapping was 16KiB with ~4.5KiB left. See the comment on the heap
value before lowering it; 192MiB was tried and made malloc hand out unmapped
memory very early in InitResourceManager.

printf %z: VitaSDK's newlib vsnprintf does not implement the 'z' length
modifier - it prints the literal characters and does not consume the
argument, shifting every later argument in the call. Where a %s follows,
strlen() then walks a garbage pointer (seen on Vita3K as a stream of
"Invalid read of uint64_t at address: 0x30"). Fixed every occurrence
compiled into the Vita build.

Makefile.vita: SPDLOG_ACTIVE_LEVEL was 6 (=OFF), so every SPDLOG_* call in
the engine was preprocessed away and BattleShip.log stayed empty no matter
what runtime level or flush policy was set. Now 2 (=INFO). Adds -DNDEBUG,
which is a TEMPORARY diagnostic: stb_truetype's glyph-outline consistency
assert fires while ImGui builds its default font atlas, and the assert's
own failure-printing path crashes inside _svfprintf_r. The underlying
vertex-count mismatch is still unfixed.

Also adds boot-sequence tracing and the two raw psp2dmp analyses of the
outstanding sceIoLseek32 crash.

+++ port/audio/audio_playback.cpp
@@ -219,8 +219,8 @@ extern "C" void portAudioSubmitFrame(const void *buf, int sampleCount)
     static bool sLoggedNonzero = false;
     if (sFirstSubmit) {
         sFirstSubmit = false;
-        port_log("SSB64 Audio: first synth frame submitted (sampleCount=%d bytes=%zu)\n",
-                 sampleCount, byteLen);
+        port_log("SSB64 Audio: first synth frame submitted (sampleCount=%d bytes=%u)\n",
+                 sampleCount, (unsigned int)byteLen);
     }
     if (!sLoggedNonzero) {
         const int16_t *s = reinterpret_cast<const int16_t *>(buf);
@@ -228,8 +228,8 @@ extern "C" void portAudioSubmitFrame(const void *buf, int sampleCount)
         for (size_t i = 0; i < count; i++) {
             if (s[i] != 0) {
                 sLoggedNonzero = true;
-                port_log("SSB64 Audio: first non-zero sample detected (idx=%zu v=%d)\n",
-                         i, (int)s[i]);
+                port_log("SSB64 Audio: first non-zero sample detected (idx=%u v=%d)\n",
+                         (unsigned int)i, (int)s[i]);
                 wavMaybeOpen();
                 break;
             }
